feat: added nightcore. feat: added a ban role for adding songs to the queue. feat: song queue can now be rewinded.
