+++ config/config.ini
@@ -87,6 +87,9 @@ TransactionTax = 0.15
 RollCases = 66, 90, 99, 100
 RollReturns = 0, 2, 4, 10
 
+[Music]
+BanRole = music_banned
+
 [Roles]
 Pronouns = She/Her, He/Him, They/Them
 Fields = Accounting, Agriculture, Anatomy and Cell Biology, Anthropology, Architecture, Biochemistry, Bioengineering, Biology, Bioresource Engineering, Chemical Engineering, Chemistry, Civil Engineering, Classics, cogito, Commerce, Computer Engineering, Computer Science, Computer Science/Biology, Cultural Studies, Desautels, Economics, Electrical Engineering, English, Experimental Medicine, Finance, Geography, History, Human Genetics, Indigenous Studies, International Development Studies, Jewish Studies, linguini, mac kid, Materials Engineering, Math, MBA, Mechanical Engineering, Medicine, Microbiology and Immunology, Neuroscience, Nursing, Pharmacology, Philosophy, Physical Therapy, Physics, Physiology, Political Science, Psychiatry, Psychology, Public Health, Social Work, Sociology, Software Engineering, Statistics, Theology, Urban Systems
